+++ sdk/storage/azure-storage-queue/tests/test_queue_client_async.py
@@ -415,12 +415,7 @@ class StorageQueueClientTestAsync(AsyncStorageTestCase):
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "azsdk-python-storage-queue/{} Python/{} ({})".format(
-                    VERSION,
-                    platform.python_version(),
-                    platform.platform()))
+            assert "azsdk-python-storage-queue/{}".format(VERSION) in response.http_request.headers['User-Agent']
 
         await service.get_service_properties(raw_response_hook=callback)
 
@@ -433,23 +428,19 @@ class StorageQueueClientTestAsync(AsyncStorageTestCase):
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "TestApp/v1.0 azsdk-python-storage-queue/{} Python/{} ({})".format(
+            assert ("TestApp/v1.0 azsdk-python-storage-queue/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
         await service.get_service_properties(raw_response_hook=callback)
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "TestApp/v2.0 azsdk-python-storage-queue/{} Python/{} ({})".format(
+            assert ("TestApp/v2.0 TestApp/v1.0 azsdk-python-storage-queue/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
         await service.get_service_properties(raw_response_hook=callback, user_agent="TestApp/v2.0")
 
@@ -460,15 +451,12 @@ class StorageQueueClientTestAsync(AsyncStorageTestCase):
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "azsdk-python-storage-queue/{} Python/{} ({}) customer_user_agent".format(
+            assert ("customer_user_agent azsdk-python-storage-queue/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
-        custom_headers = {'User-Agent': 'customer_user_agent'}
-        await service.get_service_properties(raw_response_hook=callback, headers=custom_headers)
+        await service.get_service_properties(raw_response_hook=callback, user_agent='customer_user_agent')
 
     @GlobalStorageAccountPreparer()
     @AsyncStorageTestCase.await_prepared_test