+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_base_client.py
@@ -19,8 +19,7 @@ from azure.core.pipeline.policies import (
     DistributedTracingPolicy,
     HttpLoggingPolicy,
 )
-from ._policies import CognitiveServicesCredentialPolicy, TextAnalyticsResponseHook
-from ._credential import TextAnalyticsApiKeyCredential
+from ._policies import CognitiveServicesCredentialPolicy, TextAnalyticsResponseHookPolicy
 from ._user_agent import USER_AGENT
 
 
@@ -36,7 +35,7 @@ class TextAnalyticsClientBase(object):
             credential_policy = BearerTokenCredentialPolicy(
                 credential, "https://cognitiveservices.azure.com/.default"
             )
-        elif isinstance(credential, TextAnalyticsApiKeyCredential):
+        elif hasattr(credential, "api_key"):
             credential_policy = CognitiveServicesCredentialPolicy(credential)
         elif credential is not None:
             raise TypeError("Unsupported credential: {}. Use an instance of TextAnalyticsApiKeyCredential "
@@ -56,7 +55,7 @@ class TextAnalyticsClientBase(object):
             config.retry_policy,
             credential_policy,
             config.logging_policy,
-            TextAnalyticsResponseHook(**kwargs),
+            TextAnalyticsResponseHookPolicy(**kwargs),
             DistributedTracingPolicy(**kwargs),
             HttpLoggingPolicy(**kwargs),
         ]